test_requirements.txt
##### Unit Test #####
pytest
pytest-mock
pytest-cov
pyyaml
##### KFP Components related packages #####
google-cloud-aiplatform==1.5.0
google-cloud-storage==1.42.3
google-cloud-pipeline-components==0.1.7
icecream==2.1.1
pandas==1.3.3
pandas-gbq==0.15.0
scikit-learn==0.23.2
xgboost==1.1.1
matplotlib==3.4.3
seaborn==0.11.2